add car component updates

+++ carshopBE/src/main/java/com/example/carshop/repositories/TestRepo.java
@@ -1,11 +1,9 @@
 package com.example.carshop.repositories;
 
-import com.example.carshop.entity.TestEntity;
+import com.example.carshop.domain.entity.TestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface TestRepo extends JpaRepository<TestEntity, BigInteger> {
+public interface TestRepo extends JpaRepository<TestEntity, Long> {
 }
